+++ src/test/kotlin/no/nav/tilleggsstonader/sak/vedtak/vedtaksperioderOversikt/DetaljertVedtaksperiodeTilsynBarnTest.kt
@@ -1,6 +1,8 @@
 package no.nav.tilleggsstonader.sak.vedtak.vedtaksperioderOversikt
 
 import no.nav.tilleggsstonader.sak.felles.domain.FaktiskMålgruppe
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.detaljerteVedtaksperioder.DetaljertVedtaksperiodeTilsynBarn
+import no.nav.tilleggsstonader.sak.vedtak.boutgifter.detaljerteVedtaksperioder.sorterOgMergeSammenhengende
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
